app/src/main/java/com/gmail/jorgegilcavazos/ballislife/Utils/RedditUtils.java
package com.gmail.jorgegilcavazos.ballislife.Utils;

public final class RedditUtils {
    private final static String TAG = "RedditUtils";

    /**
     * Parses a given /r/NBA flair into a readable friendly string.
     * @param flair usually formatted as "Flair {cssClass='Celtics1', text='The Truth'}"
     * @return friendly string, e.g. "The Truth", or empty string if flair was null or not valid.
     */
    public static String parseRedditNBAFlair(String flair) {
        final int EXPECTED_SECTIONS = 5;
        if (flair == null) {
            return "";
        }

        String[] sections = flair.split("'");
        if (sections.length == EXPECTED_SECTIONS) {
            return sections[sections.length - 2];
        }
        return "";
    }
}
